# swagger/swagger.yaml
basePath: /rest/api/1
consumes:
- application/json
definitions:
...
  Auth:
    properties:
      password:
        description: Password is the password required to access the platform for
          this repositroy.
        type: string
        x-go-name: Password
      secret:
        description: Hook secret to create a hook with on the respective platform.
        type: string
        x-go-name: Secret
      ssh:
        description: SSH private key.
        type: string
        x-go-name: SSH
      username:
        description: Username is the username required to access the platform for
          this repositroy.
        type: string
        x-go-name: Username
    required:
    - secret
    title: Auth is authentication option for a repository.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
  Command:
    properties:
      enabled:
        description: Enabled defines if this command can be executed or not.
        example: false
        type: boolean
        x-go-name: Enabled
      filename:
        description: Filename is the name of the file which holds this command.
        example: my_awesome_command
        type: string
        x-go-name: Filename
      hash:
        description: Hash is the hash of the command file.
        type: string
        x-go-name: Hash
      id:
        description: ID of the command. Generated.
        format: int64
        type: integer
        x-go-name: ID
      location:
        description: Location is where this command is located at. This is the full
          path of the containing folder.
        example: /tmp/krok-commands
        type: string
        x-go-name: Location
      name:
        description: Name of the command.
        type: string
        x-go-name: Name
      repositories:
        description: Repositories that this command can execute on.
        items:
          $ref: '#/definitions/Repository'
        type: array
        x-go-name: Repositories
      schedule:
        description: Schedule of the command.
        example: 0 * * * * // follows cron job syntax.
        type: string
        x-go-name: Schedule
    required:
    - name
    - id
    - filename
    - location
    - hash
    title: Command is a command which can be executed by Krok.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
  GitLab:
    properties:
      project_id:
        description: ProjectID is an optional ID which defines a project in Gitlab.
        format: int64
        type: integer
        x-go-name: ProjectID
    title: GitLab contains gitLab specific settings.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
  ListOptions:
    properties:
      end_date:
        description: EndDate defines a date of end to look for events. Not Inclusive.
        example: "2021-02-03"
        format: date-time
        type: string
        x-go-name: EndDate
      name:
        description: Name of the context for which this option is used.
        example: '"partialNameOfACommand"'
        type: string
        x-go-name: Name
      page:
        description: Current Page
        example: 0
        format: int64
        type: integer
        x-go-name: Page
      page_size:
        description: |-
          Items per Page

          required false
        example: 10
        format: int64
        type: integer
        x-go-name: PageSize
      starting_date:
        description: StartingDate defines a date of start to look for events. Inclusive.
        example: "2021-02-02"
        format: date-time
        type: string
        x-go-name: StartingDate
      vcs:
        description: Only list all entries for a given platform ID.
        format: int64
        type: integer
        x-go-name: VCS
    title: ListOptions provides options for List operations with additional filters.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
  Repository:
    properties:
      auth:
        $ref: '#/definitions/Auth'
      commands:
        description: Commands contains all the commands which this repository is attached
          to.
        items:
          $ref: '#/definitions/Command'
        type: array
        x-go-name: Commands
      events:
        description: 'TODO: Think about storing this'
        items:
          type: string
        type: array
        x-go-name: Events
      git_lab:
        $ref: '#/definitions/GitLab'
      id:
        description: ID of the repository. Auto-generated.
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: Name of the repository.
        type: string
        x-go-name: Name
      unique_url:
        description: |-
          This field is not saved in the DB but generated every time the repository
          details needs to be displayed.
        type: string
        x-go-name: UniqueURL
      url:
        description: URL of the repository.
        type: string
        x-go-name: URL
      vcs:
        description: VCS Defines which handler will be used. For values, see platforms.go.
        format: int64
        type: integer
        x-go-name: VCS
    required:
    - id
    - url
    - vcs
    - auth
    - unique_url
    title: Repository is a repository which can be managed by Krok.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
host: krok-o.io
info:
  description: Documentation the Krok API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Krok.
  version: 0.0.1
paths:
...
  /command:
    post:
      description: |-
        the command has to be edited. We don't support uploading the same thing twice.
        If the command binary needs to be updated, delete the command and upload the
        new binary.
      operationId: uploadCommand
      produces:
      - application/json
      responses:
        "201":
          description: in case of successful file upload
          schema:
            $ref: '#/definitions/Command'
        "400":
          description: invalid file format or command already exists
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to upload file, create plugin, create command or copy
            operations
          schema:
            $ref: '#/responses/Message'
      summary: Upload a command. To set up anything for the command, like schedules
        etc,
  /command/{id}:
    delete:
      operationId: deleteCommand
      parameters:
      - description: The ID of the command to delete
        format: int
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK in case the deletion was successful
        "400":
          description: in case of missing user context or invalid ID
          schema:
            $ref: '#/responses/Message'
        "500":
          description: when the deletion operation failed
          schema:
            $ref: '#/responses/Message'
      summary: Deletes given command.
    get:
      operationId: getCommand
      parameters:
      - format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/Command'
        "400":
          description: invalid command id
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to get user context
          schema:
            $ref: '#/responses/Message'
      summary: Returns a specific command.
...
  /command/add-command-rel-for-platform/{cmdid}/{repoid}:
    post:
      operationId: addCommandRelForPlatformCommand
      parameters:
      - format: int
        in: path
        name: cmdid
        required: true
        type: integer
      - format: int
        in: path
        name: repoid
        required: true
        type: integer
      responses:
        "200":
          description: successfully added relationship
        "400":
          description: invalid ids or platform not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to add command relationship to platform
          schema:
            $ref: '#/responses/Message'
      summary: Adds a connection to a platform for a command. Defines what platform
        a command supports. These commands will only be able to run for those platforms.
  /command/add-command-rel-for-repository/{cmdid}/{repoid}:
    post:
      operationId: addCommandRelForRepositoryCommand
      parameters:
      - format: int
        in: path
        name: cmdid
        required: true
        type: integer
      - format: int
        in: path
        name: repoid
        required: true
        type: integer
      responses:
        "200":
          description: successfully added relationship
        "400":
          description: invalid ids or repositroy not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to add relationship
          schema:
            $ref: '#/responses/Message'
      summary: Add a connection to a repository. This will make this command to be
        executed for events for that repository.
  /command/remove-command-rel-for-platform/{cmdid}/{repoid}:
    post:
      operationId: removeCommandRelForPlatformCommand
      parameters:
      - format: int
        in: path
        name: cmdid
        required: true
        type: integer
      - format: int
        in: path
        name: repoid
        required: true
        type: integer
      responses:
        "200":
          description: successfully removed relationship
        "400":
          description: invalid ids or platform not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to add relationship
          schema:
            $ref: '#/responses/Message'
      summary: Remove a relationship to a platform. This command will no longer be
        running for that platform events.
  /command/remove-command-rel-for-repository/{cmdid}/{repoid}:
    post:
      operationId: removeCommandRelForRepositoryCommand
      parameters:
      - format: int
        in: path
        name: cmdid
        required: true
        type: integer
      - format: int
        in: path
        name: repoid
        required: true
        type: integer
      responses:
        "200":
          description: successfully removed relationship
        "400":
          description: invalid ids or repositroy not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to add relationship
          schema:
            $ref: '#/responses/Message'
      summary: Remove a relationship to a repository. This command will no longer
        be running for that repository events.
...
  /command/update:
    post:
      consumes:
      - application/json
      operationId: updateCommand
      parameters:
      - in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/Command'
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated command
          schema:
            $ref: '#/definitions/Command'
        "400":
          description: binding error
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to update the command
          schema:
            $ref: '#/responses/Message'
      summary: Updates a given command.
  /commands:
    post:
      description: List commands
      operationId: listCommands
      parameters:
      - in: body
        name: listOptions
        schema:
          $ref: '#/definitions/ListOptions'
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            items:
              $ref: '#/definitions/Command'
            type: array
        "500":
          description: failed to get user context
          schema:
            $ref: '#/responses/Message'
...
produces:
- application/json
responses:
  Message:
    description: Message represents an error message.
...
schemes:
- http
swagger: "2.0"